Make predictor function accept and return lists
Tweak API for built predictor coroutine function (and hence eventually
generated JSON API) to accept and return lists of pydantic objects. This
enables multiple predictions in a single HTTP POST.

Given the requirement to create a pandas DataFrame for each request - this
is the only guaranteed supported signature for a pyfunc flavour mlflow model -
it is likely to be significantly more efficient to request multiple
predictions in a single HTTP request than making multiple requests.